requirements.txt
python-dotenv==0.15.0
discord.py==1.7.3
asyncio==3.4.3
discord-py-slash-command==2.4.1